chore: move to the postgrest namespace
