+++ docs/api/client.rst
@@ -5,10 +5,10 @@ To run any queries, the first step is to construct a client.
 
 The library offers both synchronous and asynchronous clients.
 
-.. autoclass:: postgrest_py.AsyncPostgrestClient
+.. autoclass:: postgrest.AsyncPostgrestClient
     :members:
     :inherited-members:
 
-.. autoclass:: postgrest_py.SyncPostgrestClient
+.. autoclass:: postgrest.SyncPostgrestClient
     :members:
     :inherited-members:
