+++ docs/conf.py
@@ -16,10 +16,10 @@
 
 
 # -- Project information -----------------------------------------------------
-import postgrest_py
+import postgrest
 
 project = "postgrest-py"
-version = postgrest_py.__version__
+version = postgrest.__version__
 release = version
 copyright = (
     "2022, Anand Krishna, Daniel Reinón García, Joel Lee, Leynier Gutiérrez González"
